Add global __metadata function
When `emitDecoratorMetadata` compiler option is set to true in `tsconfig.json`, TypeScript generates some calls to `__metadata` function.
By default in our NativeScript TypeScript projects the `noEmitHelpers` option is set to true, so TypeScript compiler does not generate the `__metadata` function.
So the application crashes at runtime. Add the missing method, so we can have both `noEmitHelpers` and `emitDecoratorMetadata` set to true in `tsconfig.json`.