fix(animations): check if target is present before removing its animation (#4586)
In Angular, if the NativeScriptAnimationModule is imported in another
NgModule more than once (which shouldn't be done), the renderer is instantiated twice. This
causes animation with empty targets to be created. If such animation is
removed, the app will crash. Additional check if the target is present
will prevent this.